requirements.txt
google-generativeai
python-dotenv
streamlit
watchdog
markdown
aiohttp
